Mailbox property pointer changes to long type.

+++ bootstrap/src/main/java/haiku/bcm/MailboxProperty.java
@@ -12,7 +12,7 @@ import haiku.vm.NativeCFunction;
  * for a given {@link MAILBOX_TAG}.
  */
 public class MailboxProperty {
-  private final int addr;
+  private final long addr;
 
   /**
    * Constructor receiving the address of the internal C rpi_mailbox_property_t struct.
@@ -20,12 +20,12 @@ public class MailboxProperty {
    * method can be called {@link Mailbox.PropertyGet()} to fetch an instance of this
    * class.
    */
-  public MailboxProperty(int addr) {
+  public MailboxProperty(long addr) {
     this.addr = addr;
   }
 
   @NativeCFunction(cImpl = "return ((rpi_mailbox_property_t *)arg1)->tag;")
-  private native int _getTag(int addr);
+  private native int _getTag(long addr);
 
   /**
    * Get the {@link MAILBOX_TAG} for the property retrieved.
@@ -36,7 +36,7 @@ public class MailboxProperty {
   }
 
   @NativeCFunction(cImpl = "return ((rpi_mailbox_property_t *)arg1)->byte_length;")
-  private native int _getByteLength(int addr);
+  private native int _getByteLength(long addr);
 
   /**
    * Get the byte length of data buffer returned for properties that return data.
@@ -47,7 +47,7 @@ public class MailboxProperty {
   }
 
   @NativeCFunction(cImpl = "return ((rpi_mailbox_property_t *)arg1)->data.value_32;")
-  private native int _getValue32(int addr);
+  private native int _getValue32(long addr);
 
   /**
    * Get the data value for properties that simply fetch an int.
@@ -59,7 +59,7 @@ public class MailboxProperty {
   }
 
   @NativeCFunction(cImpl = "return ((rpi_mailbox_property_t *)arg1)->data.buffer_8[arg2];")
-  private native byte _getBuffer8(int addr, int offset);
+  private native byte _getBuffer8(long addr, int offset);
 
   /**
    * Get the data values for properties that fetch a byte array.
@@ -72,7 +72,7 @@ public class MailboxProperty {
   }
 
   @NativeCFunction(cImpl = "return ((rpi_mailbox_property_t *)arg1)->data.buffer_32[arg2];")
-  private native int _getBuffer32(int addr, int offset);
+  private native int _getBuffer32(long addr, int offset);
   
   /**
    * Get the data values for properties that fetch a int array.
